machinetask/machinetask/requirements.txt

Django==4.0.2

djangorestframework==3.13.1